Add SIGINT handler so we can shutdown cleanly under Docker.
Normally a process doesn't have to catch SIGINT (what you get when you
send ^C) to be killed as the Linux kernel handles this for us. This
isn't true when you are running as PID 1 -- i.e. when you are inside a
docker container.

Without this fix we 1) don't shut down in a timely fashion, we have to
wait for the docker kill timeout, 2) don't give in flight requests a
chance to terminate either.

This commit fixes both of these things.